requirements.txt
# LLM 引擎
vllm>=0.6.0
pydantic>=2.0.0
pyyaml>=6.0
openai>=1.0.0
jsonschema>=4.0.0
python-dotenv>=1.0.0

# OCR 引擎
rapidocr-onnxruntime>=1.3.0
pillow>=10.0.0

# 依赖版本约束
# 注意: vllm 要求 numpy<2.0, opencv-python 更新版本需要 numpy>=2
# 如果只使用 OCR 功能，可以升级到 numpy 2.x
numpy>=1.19.5,<2.0.0
